transition to new biomes
